added modifications to main files for seccion

+++ admin/seccion.class.php
@@ -33,11 +33,13 @@
     function delete ($id){
         $this -> conexion();
         $result = [];
-        $sql = "delete from seccion where id_seccion=:id_seccion;";
-        $eliminar = $this->con->prepare($sql);
-        $eliminar -> bindParam(':id_seccion', $id, PDO::PARAM_INT);
-        $eliminar -> execute();
-        $result = $eliminar -> rowCount();
+        if (is_numeric($id)) {
+            $sql = "delete from seccion where id_seccion=:id_seccion;";
+            $eliminar = $this->con->prepare($sql);
+            $eliminar -> bindParam(':id_seccion', $id, PDO::PARAM_INT);
+            $eliminar -> execute();
+            $result = $eliminar -> rowCount();
+        }
         return $result;
     }
